add / delete / confirm / cancel buttons

+++ app/Anywakey_MacOS/Core/Home/HomeView.swift
@@ -11,6 +11,12 @@ struct HomeView: View {
     @State private var status: Bool = false
     @State private var ping: Double = 0
     
+    @State private var hoverAdd: Bool = false
+    @State private var hoverDelete: Bool = false
+    @State private var hoverCancel: Bool = false
+    
+    @State private var showDeleteConfirm: Bool = false
+    
     @State var timer: Timer?
     
     var body: some View {
@@ -30,15 +36,27 @@ struct HomeView: View {
             selectedDevice = dataService.allDevices.isEmpty ? nil : dataService.allDevices[0]
         }
         .onChange(of: selectedDevice) {
+            
+            withAnimation(.spring(duration: 0.3)) {
+                showDeleteConfirm = false
+            }
+            
             getStatus()
             
             // restart timer
             timer?.invalidate()
             
             // if address is valid create timer
-            guard selectedDevice!.BroadcastAddr.isValidAddress() else { return }
-            timer = Timer.scheduledTimer(withTimeInterval: 5, repeats: true) { _ in
-                getStatus()
+            guard selectedDevice != nil else { return }
+            if selectedDevice!.BroadcastAddr.isValidAddress() {
+                timer = Timer.scheduledTimer(withTimeInterval: 5, repeats: true) { _ in
+                    getStatus()
+                }
+            }
+        }
+        .onTapGesture {
+            withAnimation(.spring(duration: 0.3)) {
+                showDeleteConfirm = false
             }
         }
     }
@@ -118,6 +136,10 @@ extension HomeView {
             HStack {
                 addButton
                 deleteButton
+                if showDeleteConfirm {
+                    cancelButton
+                        .transition(.move(edge: .leading).combined(with: .opacity))
+                }
                 Spacer()
             }
             .padding(8)
@@ -173,25 +195,26 @@ extension HomeView {
         VStack(alignment: .leading) {
             Text("Status:")
                 .fontWeight(.semibold)
-            Text(status ? "online" : "offline")
-                .foregroundStyle(.secondary)
-                .overlay(alignment: .trailing) {
-                    Circle()
-                        .foregroundStyle(status ? .green.opacity(0.2) : .red.opacity(0.2))
-                        .overlay {
-                            Circle()
-                                .strokeBorder(lineWidth: 1)
-                                .foregroundStyle(status ? .green : .red)
-                            
-                        }
-                        .frame(height: 11)
-                        .offset(x: 17, y: 1)
-                }
+                Text(status ? "online" : "offline")
+                    .foregroundStyle(.secondary)
         }
         .padding(8)
         .frame(maxWidth: .infinity, alignment: .leading)
         .background(.secondary.opacity(0.1))
         .clipShape(RoundedRectangle(cornerRadius: 10))
+        .overlay(alignment: .trailing) {
+            Circle()
+                .foregroundStyle(status ? .green.opacity(0.2) : .red.opacity(0.2))
+                .overlay {
+                    Circle()
+                        .strokeBorder(lineWidth: 1)
+                        .foregroundStyle(status ? .green : .red)
+                    
+                }
+                .frame(width: 12, height: 12)
+                .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topTrailing)
+                .padding(6)
+        }
     }
     
     // Ping section
@@ -199,7 +222,7 @@ extension HomeView {
         VStack(alignment: .leading) {
             Text("Ping:")
                 .fontWeight(.semibold)
-            Text(ping.pingAsString())
+            Text(status ? ping.pingAsString() : " - ms")
                 .contentTransition(.numericText())
                 .foregroundStyle(.secondary)
         }
@@ -213,15 +236,24 @@ extension HomeView {
     private var addButton: some View {
         Button {
             withAnimation(.spring(duration: 0.3)) {
+                showDeleteConfirm = false
                 showAddView = true
             }
         } label: {
             Image(systemName: "plus")
-                .frame(width: 30, height: 20)
-                .background(.secondary.opacity(0.2))
+                .foregroundStyle(hoverAdd ? .white : .secondary)
+                .padding(.horizontal, 8)
+                .frame(height: 20)
+                .background(Color.gray.opacity(hoverAdd ? 0.4 : 0.2))
                 .clipShape(RoundedRectangle(cornerRadius: 5))
         }
         .buttonStyle(.borderless)
+        .scaleEffect(hoverAdd ? 1.1 : 1.0)
+        .onHover { hover in
+            withAnimation(.spring(duration: 0.3)) {
+                hoverAdd = hover
+            }
+        }
     }
     
     // Delete button
@@ -230,28 +262,75 @@ extension HomeView {
             guard selectedDevice != nil else { return }
             let index = dataService.allDevices.firstIndex(of: selectedDevice!)
             withAnimation(.spring(duration: 0.3)) {
-                dataService.delete(device: selectedDevice!)
-            }
-            if dataService.allDevices.isEmpty {
-                withAnimation(.spring(duration: 0.3)) {
-                    selectedDevice = nil
-                }
-            } else {
-                if index == 0 {
-                    withAnimation(.spring(duration: 0.3)) {
-                        selectedDevice = dataService.allDevices[index!]
+                showDeleteConfirm.toggle()
+                if !showDeleteConfirm {
+                    dataService.delete(device: selectedDevice!)
+                    if dataService.allDevices.isEmpty {
+                        withAnimation(.spring(duration: 0.3)) {
+                            selectedDevice = nil
+                        }
+                    } else {
+                        if index == 0 {
+                            withAnimation(.spring(duration: 0.3)) {
+                                selectedDevice = dataService.allDevices[index!]
+                            }
+                        } else {
+                            withAnimation(.spring(duration: 0.3)) {
+                                selectedDevice = dataService.allDevices[index! - 1]}
+                        }
                     }
-                } else {
-                    withAnimation(.spring(duration: 0.3)) {
-                        selectedDevice = dataService.allDevices[index! - 1]}
                 }
+            }
+            
+        } label: {
+            HStack {
+                if showDeleteConfirm {
+                    Text("delete")
+                        .foregroundStyle(hoverDelete ? .white : .secondary)
+                        .contentTransition(.numericText())
+//                        .transition(.move(edge: .leading).combined(with: .opacity))
+                        .frame(width: 60 ,height: 20)
+                        .background(.red.opacity(hoverDelete ? 0.4 : 0.2))
+                    
+                    
+                } else {
+                    Image(systemName: "minus")
+                        .foregroundStyle(hoverDelete ? .white : .secondary)
+//                        .transition(.move(edge: .leading).combined(with: .opacity))
+                        .frame(width: 30, height: 20)
+                        .background(.gray.opacity(hoverDelete ? 0.4 : 0.2))
+                }
+            }
+            
+            .clipShape(RoundedRectangle(cornerRadius: 5))
+        }
+        .buttonStyle(.borderless)
+        .scaleEffect(hoverDelete ? 1.1 : 1.0)
+        .onHover { hover in
+            withAnimation(.spring(duration: 0.3)) {
+                hoverDelete = hover
+            }
+        }
+    }
+    
+    private var cancelButton: some View {
+        Button {
+            withAnimation(.spring(duration: 0.3)) {
+                showDeleteConfirm = false
             }
         } label: {
-            Image(systemName: "minus")
-                .frame(width: 30, height: 20)
-                .background(.secondary.opacity(0.2))
+            Text("cancel")
+                .foregroundStyle(hoverCancel ? .white : .secondary)
+                .frame(width: 60 ,height: 20)
+                .background(.blue.opacity(hoverCancel ? 0.4 : 0.2))
                 .clipShape(RoundedRectangle(cornerRadius: 5))
         }
         .buttonStyle(.borderless)
+        .scaleEffect(hoverCancel ? 1.1 : 1.0)
+        .onHover { hover in
+            withAnimation(.spring(duration: 0.3)) {
+                hoverCancel = hover
+            }
+        }
     }
 }
